requirements.txt
# 这里列出部分关键依赖，根据实际需要可自行增删
# Below are some key dependencies, add/remove as needed

gradio~=5.5.0
langchain~=0.3.7
openai==0.27.0
sentence-transformers~=3.2.1
faiss-cpu==1.7.3
chromadb==0.3.26
tqdm~=4.66.6
numpy~=1.26.4
pandas==2.0.3
PyMuPDF==1.22.5
pdf2image~=1.17.0
pillow~=11.0.0
python-docx~=1.1.2
chardet==5.2.0
# 若需OCR支持，可引入下列依赖
# If OCR support is needed, add the below lines
pytesseract~=0.3.13
rapidocr-onnxruntime==1.3.1 
chatchat~=0.1.7
langchain-community~=0.3.5
langchain-core~=0.3.15
torch~=2.5.1
fitz~=0.0.1.dev2
transformers~=4.46.1
opencv-python~=4.10.0.84
requests~=2.32.3
pydantic~=2.9.2